seperate company from candidate component
