+++ client/src/App.js
@@ -21,7 +21,10 @@ class App extends React.Component {
   render() {
     return (
       <div>
-        <Route path="/discover" render={props => <DiscoverPage {...props} />} />
+        <Route
+          path="/discover/:type"
+          render={props => <DiscoverPage {...props} />}
+        />
         <Route exact path="/" render={props => <Landing {...props} />} />
         <Route exact path="/login" render={props => <Login {...props} />} />
         <Route
